Исправить техошибку в названии message

+++ controllers/cards.js
@@ -5,9 +5,9 @@ module.exports.createCard = (req, res) => {
     .then(card => res.send(card))
     .catch(err => {
       if(err.name === 'ValidationError') {
-        return res.status(400).send({messege: `${err.name}: 'Введены неверные данные'`});
+        return res.status(400).send({message: `${err.name}: 'Введены неверные данные'`});
       } else {
-        return res.status(500).send({messege: 'Произошла ошибка'});
+        return res.status(500).send({message: 'Произошла ошибка'});
       }
     });
 };
@@ -19,16 +19,22 @@ module.exports.getCards = (req, res) => {
 module.exports.deleteCard = (req, res) => {
   const owner = req.user._id;
   Card.findById(req.params.cardId)
-    .then((card) => {
-      if (card.owner.toString() !== owner) {
-        throw new AccessError('Нет прав на удаление карточки')
+    .then((card)=> {
+      if(card === null) {
+        throw new Error('Карточка не найдена');
+      }
+      if(card.owner.toString() !== owner) {
+        throw new Error('Нет прав на удаление карточки');
       }
       return Card.findByIdAndRemove(req.params.cardId)
         .populate(['owner', 'likes'])
         .then(card => res.send({ data: card }))
         .catch(err => res.status(500).send({message: 'Произошла ошибка'}));
     })
-    .catch(err => res.status(500).send({message: 'Произошла ошибка'}));
+    .catch(err => {
+      console.log(err.massage);
+      res.status(500).send({message: 'Произошла ошибка'})
+    });
 };
 module.exports.likeCard = (req, res) => {
   Card.findByIdAndUpdate(
